+++ VirtualFileSystem/inc/commandParser.h
@@ -1,0 +1,6 @@
+#ifndef VFS_COMMANDPARSER_H
+#define VFS_COMMANDPARSER_H
+
+void runCli(void);
+
+#endif
